Mere all resources into one stack
There are dependencies between all the stacks, better to have
all infrastrucutre hard connected with each other in same satck
